requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pyodbc==5.0.1
pandas==2.1.3
sqlalchemy==2.0.23
python-dotenv==1.0.0
boto3==1.29.7
pydantic==2.5.0


